HHH-5693 Revert back to a single compile, but without -s option
When using the two compile steps the package-info.class are getting corrupted. For this reason I am switching back to a single compile, but without the -s option to avoid the problem on Mac. As a result the meta model classes are generated in target/classes/test. On the other hand the build script is simpler, because it relies on more defaults.