HHH-5642 Error configuring Hibernate with PostgreSQL JDBC3 8.2.511 driver
The functional test introduces a test dependency to Byteman to simulate broken drivers.
The rules for Byteman are defined in testsuite/src/test/resources/bytemanrules.txt

The used rule does not work on H2 as the method signature doesn't permit it to
throw a SQLException: but as it doesn't, it can't ever have this bug of course.
So drivers which could actually be affected by this would fail as Byteman forces them
to, and the test will be able to verify Hibernate is still able to boot on those
incomplete driver implementations.