Initial discussion idea on how to pass the Audited entity to the Listener.
I know that this commit breaks public API's, and will need to be
refactored to either create a new API or some sort of sub-interface.

But I wanted to commit this to get discussion and feedback on this
approach and why/whynot this would work.

I also need to build a working copy of this so that I can test that this
sort of fix will actually help us with our issue in our product.