HHH-7239
Adding unit tests for the transaction coordinator when using shared transaction context.
One test to show that Sessions are leaked to the transaction observer, one to show that afterTransactionCompletion is not called on any session other than the main and finally one to show that the original session cannot be reused if child sessions were autoclose or flushBeforeCompletion.