HHH-8111 AttributeConverters were not mapping to the correct database types
Given an AttributeConverter of property type X and database type Y, Hibernate was attempting to map Y to the database column. This was not correct, because if Y could map to the database column the AttributeConverter wouldn't be necessary. Hibernate should have been attempting to map X, the database type that the converter converts the property to.

This only manifested if the property type was already one that Hibernate already supports, such as String, long, etc. Importantly, since Hibernate automatically handles all Serializable properties by serializing them, this problem applied to a broad set of types. If you removed Serializable from the type, for example, Hibernate would start mapping it correctly.

This VERY simple commit resolves this issue by ensuring that Hibernate maps X to the table, not Y.