HHH-8111 AttributeConverters were not mapping to the correct database types
After further experimentation and testing, I eventually discovered that the AttributeConverter work was simply never finished and never tested. I have it as close to finished as it can be, though there might be some things that the Hibernate team will want to clean up. I believe this now covers most use cases and it appears to now be working for all of my use cases.

This pull request is no longer as simple as previous commits suggested.