HHH-8623 Cache should be up to date after region eviction
* Region clear now happens within the transaction of the caller,
if any. Otherwise, a new transaction is started in order to do the
clear within a transaction and so deal with situations where cache
statistics are queried outside of a transaction.
* Cache updates after the region eviction should be allowed to
happen, so if region eviction happened within the transaction, a
putFromLoad() is mapped to a normal put instead of a PFER call,
so that the data is accessible for the current transaction. This is
not an issue for situations where region has not evicted because
the session cache will have data that's been accessed in the
transaction.
* Transaction manager could be null, if region non-transactional