HHH-10863 Be consistent in the parameter passed to ImplicitNamingStrategy#determineBasicColumnName
When dealing with an element collection (say items.name),
ImplicitNamingStrategy#determineBasicColumnName is sometimes called
with items.collection&&element.name and sometimes with items.name.

This is due to HHH-6005 which removes the "collection&&element."
part before calling determineBasicColumnName.

With this patch, we only remove the "collection&&element." part before
calling the physical naming strategy, thus allowing us to be consistent
in the way we call ImplicitNamingStrategy#determineBasicColumnName.