fix bug HHH-10256:
the problem was, that the type is checked with
statement.isWrapperFor(type), which returns true because the wrapped
class is of type CallableStatement. In the next step, the checked class
(the statement or wrapper) is casted to CallableStatement resulting in a
ClassCastException if the wrapper doesn't implement CallableStatement. I
got the ClassCastException when trying to use the SAP HANA jdbc driver
because the inner class was a CallableStatement, but the wrapper was
not.