HHH-11788 - Unignore annotations.target test dir
The current project `.gitignore` file ignores `/*/target`, which
inadvertently ignores the test package
`org.hibernate.test.annotations.target`. The test files are checked-in
despite the gitignore, so some tools complain about editing ignored
files, and any new tests might not get checked in.

This change adds a `.gitignore` to the annotations package directory,
instead of the root project gitignore, because the package dir is
a more relevant context.