HHH-6382: Allow to use @OnDelete annotation on unidirectional @OneToMany associations
Motivation for this change is that since JPA 2.x it is a valid configuration to have unidirectional one-to-many
associations i.e. do not create many-to-one 'reverse' link back to parent.

However in order to allow DDL schema export to confiure foreign  keys properly
and include ON DELETE CASCADE (for the supported dialects only) having many-to-one link is still required.

This change allow to add @OnDelete for @OneToMany association like this:

@OneToMany(cascade = CascadeType.ALL)
@JoinColumn(name = "parent_id")
@OnDelete(action = OnDeleteAction.CASCADE)
List<Child> children;

This will automatically generate ON CASCADE DELETE on table
for the entity Child.

Backward compatibility: as this change enable new
configuration which wasn't allowed before it is backward
compatible.

The only affected area is automatic DDL schema generation.