HHH-12133 - Create ManagedBeanRegistry and ManagedBean
- Do not call Instance#destroy on bean instances produced by the
fallback bean producer.
We used to, because we expected the Instance resolution to fail when a
bean could not be found. But Instance resolution will never fail in that
case, it will just return an Instance whose #isUnsatisfied() method
returns true. It happens that calling Instance#destroy did not fail with
Weld, because Weld just ignores the call in that case, but other CDI
implementations may behave differently: the javadoc, and probably also
the CDI spec, are not explicit about what happens in that case.