HHH-12866 Only generate `for update of` when necessary
We found a few test cases started to fail with previous commits. It's
caused by a combination of pagination and lock options.

Obviously in order to generate the SQL `for update of` it will needs to
put table alias. However, when the code is using `setMaxResults` to do
pagination, internally Oracle dialects will call `LIMIT_HANDLER` to
generate extra bits in the SQL which uses `rownum`. Now, because it
wraps the original SQL, none of the table aliases will be exposed. Hence
eventually the generated SQL got rejected by Oracle due to ORA-00904
(invalid identifier).

This fix is to only apply `for update of` when it's really necessary.
Thanks to `QueryLoader.applyLocks()` where it actually adds all the
`sqlAlias` with `effectiveLockMode` to the `LockOptions`. We could
compare the tables to determine whether it's *necessary* to add that `of
[tableAlias].[column]` part.