HHH-12979 - Setting hibernate.javax.cache.uri property value as relative path causes an error
Resolve the hibernate.javax.cache.uri using the ClassLoaderService.

Strip the classpath:// prefix if not resolved by new URL() in ClassLoaderService.
This way, if a framework (e.g. Spring) has installed a handler for classpath://, the resource is resolved with its handler and class loader.
If not, we remove the classpath:// prefix and we resolve the resource with our classloader.