HHH-13138 By default, pass the class loader of the test to the EMF
Not doing it causes issues when using the BytecodeEnhancerRunner which
introduces an enhancing class loader.

We could do it on a per test basis but it's easier to do it once and for
all.

And it can still be overridden anyway.

(cherry picked from commit bae98ffaccf55c78bb1eb39b1e249f9b8ae92726)