HHH-13059 Partially revert HHH-12594
The first commit was on the safe side, we decided to go the extra mile
and that was a mistake as we missed all the consequences.

The new issue is about having a shared ReaderCollector: we add the info
there for each batch which leads to collecting the elements several
times.

This reverts commit a19fc84320d1ba2172ecb37f5c8c9cbcde616712.

HHH-13059 : Correct Javadoc
(cherry picked from commit e6286e04f4f866bb555dd06dedb40a76109442f9)