HHH-13016 HHH-13199 Force parameter binding for enum literals
Currently, we don't have a way to properly handle enum literals so get back to
using parameter binding, which was used before in the "case when" case
due to a bug fixed in HHH-13001.

Note that this forces the usage of enum literals even at the root of the
select clause.