add additional target types for cast() function
- Binary
- BigDecimal, BigInteger

Also allow length or precision/scale to be specified for target types.

Fix some problems with the SQL type mappings for cast() target types.
By side-effect, fix the type of columns for BigInteger properties, which
used to have scale 2 instead of 0.

Some minor refactoring to clean up implementation of cast().