sort out the type mappings for JDBC types binary/varbinary
The correct ANSI default should be VARBINARY, not BIT VARYING, and
changing it means fewer Dialects need to override it. (Though a couple
still need to use BIT VARYING.)

There is still some inconsistency here with LONGVARBINARY/LONGVARCHAR
sometimes getting interpreted as BLOB/CLOB when not available, and
sometimes not, depending on the Dialect.