add hibernate.dialect.show_sql_functions to log function registrations
Makes it much easier for the user to get a list of all the functions the
Dialect knows about, along with their signatures.

Always show the actual signatures one uses to call the functions in HQL.

Fix the useParenthesesWhenNoArgs stuff in the function template builders.

More changes to SQL function registration in order to get a more complete
SQL function list output.

And re-allow str() and position() functions to be overridden in Dialects
by side-effect.

Also fix some very very minorly wrong function registrations.

On the other branch these are no longer synonyms, so not needed now.