clean up limit handlers
- remove all deprecated methods and duplicated code from Dialects
- refactor LimitHandlers removing even more code duplication
- add new RowsLimitHandler for Firebird/Interbase/TimesTen
- remove dirty legacyLimitHandlerEnabled stuff
- generally made everything much cleaner and simpler

For dialects which don't support an offset I've reverted to the old
pre-Hibernate-5 behavior where setFirstResult() didn't throw an
exception. The old behavior was perfectly fine and correct. If I
have setMaxResults(20) and setFirstResult(40), i.e. the 3rd page
of 20 results each, I get a perfectly well-performing

    select top 60 ....

It's wrong to complain that this query has bad performance, since
(a) it is fast enough, and (b) it is the fastest thing that you
can possibly achieve in a DB with no support for offsets or row
numbers. This change should never have been made in Hibernate 5.