BOOLEAN should map to smallint on Sybase ASE
There used to be a comment in the code explaining why, so that nobody
woudl make the mistake of mapping it to BIT, but someone removed the
comment (!), and, sure enough, I quickly checked the docs, and thought I
could just map it to BIT like on the other Transact-SQL variants, and
went ahead and changed it!

Again: the issue is that BIT columns in ASE can't be null, and can't
have indices, which doesn't match what Hibernate expects.