put back LegacyLimitHandler for now
It might be a mistake to remove this functionality, since I guess we
don't want to break people's custom Dialects without a really good
reason to do so. It's not like this is difficult code to maintain.

Nevertheless, it was correct to remove all the code that overrode
these methods in our standard Dialects. I have no idea why that code
we still there when AFAICT it was never even being called.