get date/time/duration arithmetic working on SQL Server
This involved inventing a bunch of machinery to give each Dialect
its own NATIVE basic duration resolution.

This NATIVE resolution is set to microsecond for SQL Server in
order to reduce arithmetic overflows when working with nanos. But
on other platforms we should also customize it to to either
millisecond or microsecond, depending on what maximum precision is
supported by the given platform.