add check constraints to boolean-mapped columns in Ora + Sybase
This is not 100% correct-correct, since it's vaguely possible that
someone has a converter on a Boolean property mapped to a numeric
column, and is using some other encoding than f=0, t=1. But it
seems like in those cases you wouldn't be using our DDL generation
and anyway it can be overridden in the mapping.