HHH-13500 Fix audit strategy sub-query return incorrect results
In order to maintain backward compatibility with long-standing behavior,
this introduces a new configuration option which can be toggled to have
AuditReader#find implementations adhere to returning an exact match on
revision-number rather than one which is less-than or equal-to the
provided argument.

So a new configuration option org.hibernate.envers.find_by_revision_exact_match
provides users with the ability to be able to force this new behavior
while allowing legacy behavior to remain the default.