HHH-13715 - working support for "multi-table" HQL/Criteria UPDATE and DELETE queries;
local-temp-table support works tested on H2.  I believe cte support works as well - need to set up and environment with a database supporting it.

There is a problem saving entities with secondary tables atm so for now i cannot create the data to assert that the correct rows were deleted.  But looking at the executed SQL it works