HHH-13845 - Work for refactor in StatelessSessionImpl for unused method parameter.
StatelessSessionImpl has a public method afterOperation(boolean success). The method parameter 'success' was used and passed as a parameter in afterNonTransactionalQuery().

Since the change was done long time back, the success variable is not being used now, and being passed by its caller in the same class now. In order to avoid that, a refactor can be done.

Since it is a public method, this might have an impact as well. But if not, this is fixed in this pull request.