HHH-13852 - Work for method renaming in BaseCoreFunctionalTestCast
Currently, in BaseCoreFunctionalTestCase a method named getXmlFiles()has a TODO placed that suggests the method name to be changed to getOrmXmlFiles().

Since the method is a protected methods, there is a possibility it is being used by other clients. Hence, changing the method name in minor versions might break the code for clients.

Although, this long-awaited change can be done as an improvement in wip/v6.0