move several new types out of the root package
The namespace org.hibernate is already much too cluttered and
we should try hard to not make it worse in H6.

Furthermore, some of these things look to be of more interest
to developers of Hibernate than to users.

Now, I'm not sure that org.hibernate.query is absolutely the
best place either, but it's at least better.