HHH-14589 - Make sure documentation examples regarding basic-types work and update the section
- re-enable tests
- re-organize some tests
- fix basic-type resolution in various cases from tests in both `hibernate-core` and `documentation`
- updated the basic_types.adoc user-guide section + moved much of the "legacy" discussion to an appendix

A major change here is the change to the `AdjustableBasicType` contract (previously `SqlTypeDescriptorIndicatorCapable`).  We know pass along the "real" domain type to be able to use that in the adjustment.  Mainly this was needed for better primitive/wrapper (and primitive/wrapper array) handling

The basic-types section of the user-guide has been expanded and re-organized significantly, although it is not quite complete