sort out some confusion regarding elements() vs value() + indices() vs index()
Strictly, elements() and indices() don't make sense as select items, but we
have tests for this, and users who reported bugs and sent in patches, etc, etc,
so I'm going to go ahead and keep accepting them in the select clause as a
blessed misuse.

I'm not however going to allow them to be dereferenced because no, that's why.