HHH-15466 Compatibility with Jandex 3.0.0
The only change in Jandex 3.0.0 relevant to Hibernate ORM is that
`Indexer.index()` used to return `ClassInfo`, but now returns `void`.
This is a breaking change, but Jandex 3.0.0 has a synthetic bridge
method with the old signature for binary compatibility -- except
it always return `null`. Therefore, with this commit, Hibernate ORM
simply ignores the return value completely, which makes it compatible
with Jandex 3.0.0 at runtime, even though it is still compiled against
Jandex 2.4. The code is also source-compatible with Jandex 3.0.0 if
that is ever needed.