HHH-15668 Apply recommended settings for CockroachDB
The recommended CockroachDB v22.1 settings for local testing have changed. The recommended
settings are now applied.

The setting `schemachanger.backfiller.buffer_increment` seems to have been removed
(despite still being mentioned in the documentation). We remove it
because it throws an error and this may leave the Cockroach cluster
with some of the recommended settings not applied.