HHH-15393: fix for what looks like a copy & paste error
In the changes added to BasicCollectionPersister during HHH-15393, it looks as if generateUpdateRowOperation has maybe been copy-pasted from generateInsertRowOperation (which is perfectly fine), but when the custom sql update row operation is generated, the mutation details for the insert case rather than the ones for the update case are fetched. In case that this assumption is correct, there are some more in buildCustomSqlUpdateRowOperation, where getDeleteDetails() is used instead of getUpdateDetails().