HHH-16812 Test reproducing the issue
It seems that when trying to call the Smell.setCheese()setter and when
the entity is an uninitialized Food proxy there’s a ClassCastException
(maybe the proxy should be narrowed to the actual class of the entity)
and then it tries to build an exception but to do that it needs to
initialize the entity and that fails again with the same problem.
Building the exception fails because PropertySetterAccessException tries
to call the toString method of the uninitialized proxy.