HHH-17742 Test for race condition in ConcreteSqmSelectQueryPlan
Race condition occurs when two or more concurrent reach the synchronized
block in ConcreteSqmSelectQueryPlan#withCacheableSqmInterpretation. The
latter ones will see the cacheableSqmInterpretation by the first one,
but don't check whether it is compatible
(jdbcSelect.dependsOnParameterBindings(), jdbcSelect.isCompatibleWith).

On MySQL this can cause "limit null,1" to be rendered if the first query
has both offset and limit, the latter ones only a limit.