HHH-18286 Add Setting to fail Bootstrapping on Metadata Access Failure
A failure to access the JDBC connection metadata while bootstrapping
Hibernate can lead to ClassCastExceptions during runtime when the
Oracle JDBC driver is used and a BLOB is created by the
NonContextualLobCreator. To avoid this issue, a new JDBC setting is
introduced, which lets the bootstrapping fail when the connection
metadata cannot be obtained.