HHH-13046 ClassLoaderService in JCacheRegionFactory
This enables the use of caching in scenarios where a non-standard class
loader is registered using the service registry.

Unfortunately, since the service registry lifecycle is tied to the
lifecycle of the SessionFactory, the required initialization happens
too late for some of the test cases to "just work".

To keep compatibility with the specs from the test suite, some hackery
is required: the JCacheRegionFactory has to hold on to some defaults
from before being set up.