remove additional modules which don't contain APIs from the generated Javadoc
These modules don't contain APIs which are called by users or implemented by
integrators. Including them in the generated Javadoc results in confusion
for users and makes the Javadoc much more difficult to navigate and consume.

In particular, we definitely don't want packages like these included in the
main Javadoc distribution for Hibernate ORM:

  org.hibernate.testing.orm.domain.gambit
  org.hibernate.testing.orm.domain.helpdesk

"Good documentation" isn't about having the maximum possible quantity of
documentation; it's about having documentation where people actually have a
hope of finding what they're looking for. In general, generated Javadoc is
entirely useless except when we've actually taken time to actually write
meaningful and readable prose. If all I need is a list of classes and their
methods, my IDE is quite capable of presenting that information to me.

Worse, documentation for things that aren't API appears to invite users to
interact with these things, which is precisely what we don't want them to do!

I'm leaving Envers alone, but I'm honestly not even sure if the Envers packages
belong here, since it doesn't look like anyone has actually spent much time
writing Javadoc for these packages. If Envers Javadoc is useful, it would be
better to put it in a separate Javadoc build, IMO.