give some love to org.hibernate.Transaction
also clean up some stuff in AbstractPersistentCollection

Unfortunately TransactionStatus is an SPI type, which makes
its use in the API of Transaction a bit wrong. Also, it's
weird to see dependence on an interface defined by JPA.
(However these are relatively harmless aesthetic points, so
I decided not to deprecate anything.) Instead I've added
alternative "convenience" operations.

Note on terminology: a transaction which has been "marked for
rollback only" is still an active transaction! You can still
do work in it, and that work will still be atomically rolled
back when the transaction completes. Apparently there was
some confusion on this at some point, which was later fixed,
but some evidence of the confusion was left behind in code.