HHH-18979 prototype for Restrictions
We will need an API like for this for out Jakarta Data 1.1 implementation.
But it's actually really quite generally useful to be able to add some
restrictions to a query without needing to jump through the hoops of the
JPA Criteria API. In some ways, this reminds me of the old Hibernate
criteria API from pre-JPA days (but typesafe, of course).