remove lots of unnecessary calls to getRuntimeMetamodels()
I completely understand the desire to not go to the SessionFactoryImplementor
for everything, and I'm totally onboard with that, but this isn't actually
helping at all (if there ever came a time when we needed to, we can inline
those convenience methods completely *trivially*). If we want to actually not
depend on the SFI, we need to start passing RuntimeMetamodelsImplementor around
instead. That would be *awesome* but it's probably a lot more work.

For now the value of the convenience methods is clear, it seems to me.